removes unused packages and classes
