+++ app/src/main/java/com/androidstudy/mpesa/ui/PaymentActivity.kt
@@ -18,7 +18,6 @@ package com.androidstudy.mpesa.ui
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
 import com.androidstudy.daraja.Daraja
 import com.androidstudy.daraja.callback.DarajaException
 import com.androidstudy.daraja.callback.DarajaListener
@@ -26,12 +25,9 @@ import com.androidstudy.daraja.callback.DarajaPaymentListener
 import com.androidstudy.daraja.model.AccessToken
 import com.androidstudy.daraja.model.PaymentResult
 import com.androidstudy.daraja.util.Environment
-import com.androidstudy.mpesa.Config
+import com.androidstudy.mpesa.utils.Config
 import com.androidstudy.mpesa.R
-import com.androidstudy.mpesa.common.BaseActivity
-import com.androidstudy.mpesa.common.Status
 import com.androidstudy.mpesa.utils.AppUtils
-import com.androidstudy.mpesa.viewmodel.PaymentViewModel
 import kotlinx.android.synthetic.main.activity_payment.*
 import kotlinx.android.synthetic.main.content_payment.*
